pyproject.toml
[project]
name = "grid_trading_bot"
version = "0.1.0"
description = "Open-source cryptocurrency trading bot designed to perform grid trading strategies using historical data for backtesting."
authors = [{ name = "Jordan TETE", email = "tetej171@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.12"
dependencies = [
    "pandas==2.2.3",
    "numpy==2.1.3",
    "plotly==6.3.0",
    "tabulate==0.9.0",
    "aiohttp==3.10.11",
    "apprise==1.9.5",
    "ccxt==4.4.82",
    "configparser==7.2.0",
    "psutil==7.1.3",
    "python-dotenv==1.2.1"
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==0.26.0",
    "pytest-cov==7.0.0",
    "pytest-timeout==2.4.0",
    "pre-commit==4.3.0",
]

[project.urls]
repository= "https://github.com/jordantete/grid_trading_bot"
issues = "https://github.com/jordantete/grid_trading_bot/issues"
discussions = "https://github.com/jordantete/grid_trading_bot/discussions"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
markers = [
    "asyncio: mark a test as an async test",
    "timeout: mark a test with a timeout"
]

[tool.pytest_asyncio]
mode = "auto"

[tool.coverage.run]
omit = [
    "*/interface*.py"
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["core", "strategies", "config", "utils", "tests"]
exclude = ["backups", "*.bak", "*.backup", ".venv", "venv", "__pycache__"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "DTZ",  # flake8-datetimez
    "TCH",  # flake8-type-checking
]
ignore = [
    "S101",   # assert used (OK in tests)
    "COM812", # missing trailing comma (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["core", "strategies", "config", "utils"]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
# CLI scripts intentionally use print() for console output
"adaptive_scanner.py" = ["T201"]
"family_scanner.py" = ["T201"]
"stock_trading_assistant.py" = ["T201"]
"optimization/asset_optimizer.py" = ["T201"]
"run_all_backtests.py" = ["T201"]
"run_optimization.py" = ["T201"]
"tools/cli_tools.py" = ["T201"]
"tools/health_check.py" = ["T201"]
"market_diagnostics.py" = ["T201"]
"multi_pair_runner.py" = ["T201"]
"pair_validator.py" = ["T201"]
"simple_adaptive_scanner.py" = ["T201"]
"triple_threat_trader.py" = ["T201"]
"triple_threat_trader_coinbase.py" = ["T201"]
"main.py" = ["T201"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["logs", "data"]

[tool.setuptools]
include-package-data = true
